+++ src/Recyclarr.Repo/IMultiRepoUpdater.cs
@@ -2,5 +2,5 @@ namespace Recyclarr.Repo;
 
 public interface IMultiRepoUpdater
 {
-    Task UpdateAllRepositories(CancellationToken token);
+    Task UpdateAllRepositories(CancellationToken token, bool hideConsoleOutput = false);
 }